add JNIEntryPoint.c
